PFL-13 refactoring: file name and add container
