+++ pfl-frontend/src/components/main/main.component.scss
@@ -1,10 +1,176 @@
-.header {
-    transition: opacity 0.5s ease;
-    opacity: 1;
-    z-index: 9999;
+/* main-container styles */
+div.main-container {
+    position: relative;
+    width: 100%;
+    height: 100%;
 }
 
-.header.hide {
-    opacity: 0;
-    pointer-events: none;
+/* main styles */
+div.main {
+    display: flex;
+    height: 100%;
+    overflow: scroll;
+    flex-direction: column;
+
+    // NOTE(hajae): description styles
+    > div.discription > span {
+        position: absolute;
+        left: 200px;
+        font-size: 60px;
+        font-weight: 700;
+        color: #f5f5f5;
+
+        &.discription1 {
+            top: 250px;
+        }
+
+        &.discription2 {
+            top: 350px;
+            white-space: nowrap;
+
+            > span.position {
+                color: #30475E;
+            }
+        }
+
+        &.discription3 {
+            top: 450px;
+
+            > span.name {
+                font-size: 72px;
+                text-decoration-line: underline;
+                text-decoration-color: #F05454;
+                text-decoration-thickness: 12px;
+            }
+        }
+    }
+
+    // NOTE(hajae): scroll icon styles
+    > img.plz-scroll {
+        position: fixed;
+        bottom: 0;
+        left: 42%;
+        animation: bounce 2s infinite;
+        z-index: 9999;
+        transition: opacity 0.5s ease;
+        opacity: 1;
+
+        &.hide {
+            opacity: 0;
+            pointer-events: none;
+            z-index: 0;
+        }
+    }
+
+    // NOTE(hajae): main-career styles
+    > div.main-career {
+        height: 2000px;
+        background-color: #f5f5f5;
+        overflow: hidden;
+        position: relative;
+
+        > span.main-title {
+            margin: 0;
+            padding: 0;
+            position: absolute;
+            bottom: 92%;
+            left: 20%;
+            opacity: 0;
+            white-space: nowrap;
+            color: #30475E;
+            font-size: 50px;
+            font-weight: 700;
+        }
+
+        > div.line {
+            width: 10px;
+            height: 90%;
+            background-color: #30475E;
+            position: absolute;
+            left: 30%;
+            top: 10%;
+            opacity: 0;
+            transform: translateX(-50%);
+
+            &.show {
+                animation: lineSlideIn 1s ease-in-out forwards;
+            }
+        }
+
+        // NOTE(hajae): date-container styles
+        > div.date-container {
+            position: absolute;
+            left: 12%;
+            top: 13%;
+            transform: translateX(-50%);
+            display: flex;
+            flex-direction: column;
+            align-items: center;
+
+            > div.date {
+                font-size: 36px;
+                font-weight: 900;
+                color: #F05454;
+                opacity: 0;
+                transition: opacity 2s ease;
+                margin-top: 5px;
+                margin-bottom: 400px;
+
+                &.current {
+                    opacity: 1;
+                }
+            }
+        }
+    }
+}
+
+// NOTE(hajae): animation keyframes
+@keyframes bounce {
+    0%, 20%, 50%, 80%, 100% {
+        transform: translateY(0);
+    }
+    40% {
+        transform: translateY(-20px);
+    }
+    60% {
+        transform: translateY(-10px);
+    }
+}
+
+@keyframes lineSlideIn {
+    100% {
+        left: 25%;
+        opacity: 1;
+    }
+}
+
+/* TODO(hajae): Component화 */
+div.main-career > span.main-section1 {
+    bottom: 28%;
+    text-align: center;
+    font-size: 0;
+}
+
+div.main-career > span.main-section1 > img {
+    margin-bottom: 10px;
+    display: block;
+    width: 400px;
+    border-radius: 10px;
+}
+
+div.main-career > span.main-section1 > span {
+    font-size: 24px;
+    color: #30475E;
+}
+
+/* animation styles */
+div.main-career.show span {
+    animation: slideIn 1s ease-in-out forwards;
+}
+
+@keyframes slideIn {
+    100% {
+        left: 10%;
+        opacity: 1;
+    }
 }
