+++ pfl-frontend/src/containers/main/main.container.scss
@@ -1,0 +1,10 @@
+.header {
+    transition: opacity 0.5s ease;
+    opacity: 1;
+    z-index: 9999;
+}
+
+.header.hide {
+    opacity: 0;
+    pointer-events: none;
+}
